corrected the paths after renaming
